Add cache; Update roles and permission controller; Update role permission policy; Completed role and permissions pages and functionality

+++ api/app/Http/Controllers/UserGroupController.php
@@ -58,6 +58,9 @@ class UserGroupController extends Controller
     public function destroy(UserGroup $userGroup): Response
     {
         $this->authorize('delete', $userGroup);
+        if ($userGroup->users()->exists()) {
+            return $this->error('User group is not empty', 400);
+        }
         $userGroup->deleted_by = Auth::id();
         $userGroup->save();
         $userGroup->delete();
